IoWndChangeFont() の引数は LOGFONT * でなく HFONT にする IoWnd.c の中でもフォント情報は HFONT として保持する。ioOnCreate()でSYSTEM_FIXED_FONTを取得しておき、 上位モジュールからセットされたらそのHFONT情報を使うようにする MainWnd.c ではIoWndの HFONT 情報を保持する
git-svn-id: http://mozuku/svn/editorRepo/tedd/trunk@120 dbae37e1-483b-0410-8081-86c1b21c2f05

+++ sys/WinMain/WinMain.c
@@ -49,7 +49,7 @@ WinMain( HINSTANCE hInstance, HINSTANCE hPrevInstance, PSTR szCmdLine, int nCmdS
         }
     }
 
-    return msg.wParam; /* PostQuitMessage()‚Ìˆø” */
+    return (int)msg.wParam; /* PostQuitMessage()‚Ìˆø” */
 }
